refactor: split monolithic backend into feature modules
